Fixes issue where mapbox and support vars didn't exist
