+++ bods-nextjs/app/guidance/support/bus-operators/sections/SupportEmailLink.tsx
@@ -3,12 +3,14 @@
  * Renders a linked email address using the configured support email.
  */
 
-import { config } from '@/config';
+import { useSupportConfig } from '@/app/guidance/support/SupportConfigProvider';
 
 export function SupportEmailLink() {
+  const { supportEmail } = useSupportConfig();
+
   return (
-    <a className="govuk-link" href={`mailto:${config.supportEmail}`}>
-      {config.supportEmail}
+    <a className="govuk-link" href={`mailto:${supportEmail}`}>
+      {supportEmail}
     </a>
   );
 }
