+++ bods-nextjs/app/publish/org/[orgId]/dataset/avl/new/page.tsx
@@ -1,15 +1,19 @@
 import type { Metadata } from 'next';
+import { connection } from 'next/server';
 import { ProtectedRoute } from '@/components/auth/ProtectedRoute';
+import { config } from '@/runtime-config';
 import { AVLCreatePageContent } from '../_components/AVLCreatePageContent';
 
 export const metadata: Metadata = {
   title: 'Publish new data feed',
 };
 
-export default function AVLCreatePage() {
+export default async function AVLCreatePage() {
+  await connection();
+
   return (
     <ProtectedRoute>
-      <AVLCreatePageContent />
+      <AVLCreatePageContent avlIpAllowList={config.avlIpAllowList} />
     </ProtectedRoute>
   );
 }
