--- bods-nextjs/lib/server-config.ts
@@ -1,5 +1,0 @@
-const djangoInternalOrigin = process.env.DJANGO_INTERNAL_ORIGIN || 'http://localhost:8000';
-
-export const serverConfig = {
-  djangoInternalOrigin,
-} as const;
